Move and refactor article pages
