+++ frontends/mit-open/src/pages/ArticleUpsertPages/ArticleCreatePage.test.tsx
@@ -1,13 +1,12 @@
 import React from "react"
 import {
   renderRoutesWithProviders,
-  renderWithProviders,
   screen,
   user,
   waitFor,
   within,
 } from "../../test-utils"
-import { ArticlesCreatePage, ArticleEditingPage } from "./ArticleUpsertPages"
+import ArticleEditPage from "./ArticleEditPage"
 import type { Article } from "api"
 import { articles as factory } from "api/test-utils/factories"
 import { makeRequest, setMockResponse, urls } from "api/test-utils"
@@ -20,7 +19,7 @@ describe("ArticlesEditPage", () => {
       [
         {
           path: "/articles/:id/edit",
-          element: <ArticleEditingPage />,
+          element: <ArticleEditPage />,
         },
         {
           path: "*",
@@ -124,37 +123,4 @@ describe("ArticlesEditPage", () => {
       }
     },
   )
-})
-
-describe("ArticlesCreatePage", () => {
-  test("Has 'Save' and 'Cancel' but not 'Delete' buttons", async () => {
-    renderWithProviders(<ArticlesCreatePage />)
-    await screen.findByRole("button", { name: "Save" })
-    await screen.findByRole("button", { name: "Cancel" })
-    const deleteButton = screen.queryByRole("button", { name: "Delete" })
-
-    expect(deleteButton).toBe(null)
-  })
-
-  test("Calls creation endpoint when saved", async () => {
-    const article = factory.article()
-    const { location } = renderWithProviders(<ArticlesCreatePage />)
-    const titleInput = await screen.findByLabelText(/Title/)
-    const bodyInput = screen.getByTestId("mock-ckeditor")
-    const saveButton = screen.getByRole("button", { name: "Save" })
-
-    await user.click(titleInput)
-    await user.paste(article.title)
-    await user.click(bodyInput)
-    await user.paste(article.html)
-
-    setMockResponse.post(urls.articles.list(), article)
-    await user.click(saveButton)
-    expect(makeRequest).toHaveBeenCalledWith("post", urls.articles.list(), {
-      title: article.title,
-      html: article.html,
-    })
-
-    expect(location.current.pathname).toBe(`/articles/${article.id}`)
-  })
 })
