+++ frontends/mit-open/src/pages/ArticleUpsertPages/index.ts
@@ -1,0 +1,4 @@
+import ArticleEditPage from "./ArticleEditPage"
+import ArticleCreatePage from "./ArticleCreatePage"
+
+export { ArticleEditPage, ArticleCreatePage }
